Better error support and control of the general behaviour
Added possibility for a callback to be passed when render is finished,
this add possibility for the developper to check for errors.
Added two flags which preserve the current behaviour for backward
compability :
- failSilently: true/false : When less send back an error, ignore it
(default true)
- answerRequest : true/false : Should connect-lesscss send the response
itself or
pass this control to the developper with err and the compiled css
(default true).

When answerRequest is true and render return an error then a status 500
response is sent to the server.